unify local helper files into helpers.go
